Properly encapsulate map parsing in a class - MapParser
Previously map parsing functions did not belong to a class as they would be in a C program.
Functions to access the special sections (like the stack) from the result of parsing was also added. This stops the calling code from having to iterate over the returned vector to find the stack.