UNITTEST Update unit tests to utilise new MapParser class
Outside of the map-parser-test the changes are simply creating a MapParser instance then calling ParseMap.
For the map-parser-test a #define has been introduced to make the ParseLine function public for unit tests.
The #define is undef'd straight after the #include of map-parser.hpp is complete. This change does not affect the code under /src at all. ParseLine cannot be called public ally by any other unit test file or any code under /src.
This isn't the cleanest approach as whilst generally private functions shouldn't be unit tested but ParseLine is large enough to warrant testing and doesn't make sense as it's own class.
The other option was creating a MapParserTest class which was a friend to MapParser and marking ParseLine a friend function but that seemed even more convoluted.