Improve error handling and documentation in DeepCopy
Comment added to the header outlining the responsibility the calling function accepts when utilising DeepCopy.

std::bad_alloc exceptions are now caught and handled. The problem is passed back to the caller via returning nullptr. An error message is also printed to the console.

The error message when the process_vm_readv call itself fails is also improved to include the error message from strerror()

The function has been renamed to DeepCopy from deepCopy as per Google's C++ coding style. This commit will not build because of this however a followup commit will change the function name in calling functions.