Separate RemoteHeapPointer to utils header and add ostream overload
Add utils.hpp to contain structs frequently used across most of the code base. Previously this struct was defined in a random header file.

The ostream operator added in the .cpp file is rudimentary and is just for convenient debugging

Comments added to explain the use of every member variable

The RemoteHeapPointer members need to be renamed to meet Google C++ guidelines