Don't fetch the size pointed to for pointers to the start of the heap
There is a large whitespace changes in this commit to bring the indentation inline with Google C++ guidelines. The only actual changes are on lines 64 to 76

Pointers to the very first byte of the heap would previously make the program malfunction when we attempted to get the size they pointed to via GetMallocMetadata.

This was due to reading the 8 bytes before the pointer which would not actually be part of the heap and may not belong to the process specified in the PID parameter of process_vm_readv. This would cause process_vm_readv to return a Bad Address error.

The fix is to not fetch their size via GetMallocMetadata but just assign them a size of 0.