UNITTEST Update target asm program to not rely on STDIN existing
The unit test was failing in Jenkins because it was attempting to open STDIN when the getchar() syscall was invoked. However Jenkins calls the programs from a non interactive terminal so it was failing.

The program now just waits 5 seconds which is enough time for the unit test to run