FIX not default initialising MapParser member variable
Not default initialising the heap struct would result in garbage values being returned if the program that was parsed does not have a heap. Now an struct with empty values is returned. 

This fixes the "0 pointers target program" unit test which would fail intermittently depending on the garbage values returned as the program being analysed in this test has no heap.