Improve the performance of and properly encapsulate HeapTraverser
HeapTraverser isn't just several static methods under it's namespace. The functionality of the class is the same however it's performance is considerably better at around half the runtime duration. 

Malloc metadata also had to be updated with an additional default parameter to accommodate HeapTraverser new logic. This won't affect any calling code as the default parameter value keeps it's behaviour as before this commit.

Main and unit tests are also updated to incorporate the new HeapTraverser encapsulation