Utilise every thread available to traverse the heap
Each hardware thread is used i.e. 8 threads on a 8 thread CPU is used to traverse the heap.
The threads are synchronised via a std::atomic containing the index of a vector containing the base pointers. This guarantees no threads will traverse the same base pointer.
The threads operate on the same copy of the heap which introduces a race condition in that two threads could get to a heap object at the exact same time and both see it as untraversed and both then mark is as traversed. This will lead to both threads doing the same work for a short amount of time. This race condition does NOT introduce any memory safety issues or any greater chance of false positives or incorrect traversals.

The performance increase from multithreading heap traversal is close to linear. Meaning an 8 core CPU now runs the program almost 8 times as fast. Previous tests that took 5 minutes to complete now run in 37 seconds.