UNITTEST Add helper to run given function against every target program
Allows for boilerplate to be removed from unit tests as a condition/scenario can be tested against every program by creating a lambda that does it for one function.

This also helps the maintainability of the unit tests as when a target program is added it can be added to the list of target programs and every unit test will then run on it.